refactor: major rebuild of existing code structure and fixes to patches
