+++ LethalProgression/Components/EnemyDamageTrackerComponent.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using GameNetcodeStuff;
+using UnityEngine;
+
+namespace LethalProgression.Components
+{
+    internal class EnemyDamageTrackerComponent : MonoBehaviour
+    {
+        List<PlayerControllerB> playersWhoHarmedEnemy = new List<PlayerControllerB>();
+    }
+}
