+++ LethalProgression/Network/PlayerJumpHeightData.cs
@@ -1,8 +1,14 @@
 namespace LethalProgression.Network;
 
-internal class PlayerJumpHeightData
+internal struct PlayerJumpHeightData
 {
     public ulong clientId;
 
     public int jumpSkillValue;
+
+    public PlayerJumpHeightData(ulong clientId, int jumpSkillValue)
+    {
+        this.clientId = clientId;
+        this.jumpSkillValue = jumpSkillValue;
+    }
 }
