+++ LethalProgression/Network/SaveProfileData.cs
@@ -1,16 +1,15 @@
 using LethalProgression.Saving;
 
-namespace LethalProgression.Network
-{
-    internal class SaveProfileData
-    {
-        public ulong steamId { get; set; }
-        public SaveData saveData { get; set; }
+namespace LethalProgression.Network;
 
-        public SaveProfileData(ulong steamId, SaveData saveData)
-        {
-            this.steamId = steamId;
-            this.saveData = saveData;
-        }
+internal struct SaveProfileData
+{
+    public ulong steamId;
+    public SaveData saveData;
+
+    public SaveProfileData(ulong steamId, SaveData saveData)
+    {
+        this.steamId = steamId;
+        this.saveData = saveData;
     }
 }
