+++ LethalProgression/Skills/UpgradeType.cs
@@ -1,0 +1,14 @@
+namespace LethalProgression.Skills;
+
+public enum UpgradeType
+{
+    HPRegen,
+    Stamina,
+    Battery,
+    HandSlot,
+    Value,
+    Oxygen,
+    JumpHeight,
+    SprintSpeed,
+    Strength,
+}
