+++ global.json
@@ -1,0 +1,5 @@
+{
+  "sdk": {
+    "version": "7.0.407"
+  }
+}
